consistent naming with api and helpers and consolidation, update imports
